Prior to change

+++ app/src/main/java/com/example/revoluttest/model/Currency.kt
@@ -3,4 +3,4 @@ package com.example.revoluttest.model
 class Currency(val currencyName : String,
                val countryCurrencyName: String,
                val flag: Int,
-               val value: Int)
+               val value: Float)
